+++ scheduler/conf.c
@@ -3895,10 +3895,10 @@ read_policy(cups_file_t *fp,		/* I - Configuration file */
             char        *policy,	/* I - Location name/path */
 	    int         linenum)	/* I - Current line number */
 {
-  int			i;		/* Looping var */
+  size_t			i;		/* Looping var */
   cupsd_policy_t	*pol;		/* Policy */
   cupsd_location_t	*op;		/* Policy operation */
-  int			num_ops;	/* Number of IPP operations */
+  size_t			num_ops;	/* Number of IPP operations */
   ipp_op_t		ops[100];	/* Operations */
   char			line[HTTP_MAX_BUFFER],
 					/* Line buffer */
@@ -3964,7 +3964,7 @@ read_policy(cups_file_t *fp,		/* I - Configuration file */
 	if (*valptr)
 	  *valptr++ = '\0';
 
-        if (num_ops < (int)(sizeof(ops) / sizeof(ops[0])))
+        if (num_ops < (sizeof(ops) / sizeof(ops[0])))
 	{
 	  if (!_cups_strcasecmp(value, "All"))
 	    ops[num_ops ++] = IPP_ANY_OPERATION;
